backend/requirements.txt
fastapi
uvicorn[standard]

python-dotenv
requests

groq
google-genai

deepgram-sdk
tavily-python
googlesearch-python
arxiv
pymupdf

chromadb
SQLAlchemy

numpy
num2words

python-multipart
google-generativeai
